New PluginDeps context manager (#104)
* Created new PluginDeps context manager

* Moved pandas import inside the function to correctly reload it.

* Moved pandas import inside the function to correctly reload it.

Deleted some random variables

* Added warning about importing in scoped functions

* Fixed bug that removed the whole repo if no Deps folder was detected

* Log zip contents and print plugin sucesfully isntalled

+++ Devtools/clean.sh
@@ -14,9 +14,11 @@ rm -rf HorusAPI/src/*.c
 
 # Remove NBDSuite plugin deps (except for nbdsuite)
 # in the AppSupport/Plugins/NBDSuite/deps directory
-cd AppSupport/Plugins/NBDSuite/deps
-find . -maxdepth 1 -not -name 'nbdsuite*' -not -name "." -not -name ".." -exec rm -rf {} +
-cd ../../../../
+if [ -d "AppSupport/Plugins/NBDSuite/deps" ]; then
+    cd AppSupport/Plugins/NBDSuite/deps
+    find . -maxdepth 1 -not -name 'nbdsuite*' -not -name "." -not -name ".." -exec rm -rf {} +
+    cd ../../../../
+fi
 
 # Remove the Horus plugin deps
 # in the AppSupport/DefaultPlugins/Horus/deps directory
